Notifications
This does not resolve the flicker, but does make it check against the
right status code. The issue is that removing a torrent, forces an
update, in which the notification is of course removed after the
re-render. Would need to implement an actual notification system for
this. Maybe there is a lib? Otherwise, should be pretty easy, and fun.