MP-4: Modifier wizard
Currently not possible to order modifiers or their values. I'll most
likely get around the values not being ordered by ordering them in the
page form.

But ordering modifiers should be possible. Will have to figure out how
to do so nicely. I'd really want to make it a drag and drop :)