MP-100: Cleanup for mobile
- Always print stack trace
- Don't log the woff2 primeng icons
- Logged one new bug in pages-dirs for mobile stuff

The expierence on desktop has gone up a lot, it's a lot nicer. Mobile is
fine still but the tables are too nice on it. I don't really want to
change it, since they're great on desktop (intended use anyway). But
will have to look into further at some point