Code reworks for partial downloads and future updates
This commit reworks how the content service communicates with the
download clients, and its content. Two new interfaces are introduced to
present those.
This reworks takes a lot of pressure of the ProviderAdapter, which it
shouldn't have had in the beginning. And allows the content service to
directly communicate with Content so less hops have to be made