MP-119: SignalR support
Good bit of working around fiber and fasthttp. But got it all working,
this might still all fail in the future. But I think it should be all
good now.

The logger used by the signalr lib is a weird and annoying, so got an
awkward adaptor for zerolog for it;