MP-116: Fix Pasloe content not being cancelled correctly in loading state
Adds context to a lot of function, I think this is the more correct way
to work with functions. I'm not sure if I'm supposed to do it this way
with the logging? As I don't really need to log that something got
canceled.
Context is pretty cool I think, works nicely like this!