More rate limit patching
Introduces a new httpClient with automatic rate limitting. Not using
this by default everywhere. As I'm not certain all http requests happen
on a different thread, and I don't want to sleep the entire program...
Missing Java Futures here tbh haha.

Pasloe: if the amount of images in the content is lower than 5, sleep 1.
As the sleeper would've have kicked it otherwise.